+++ app/Database/Migrations/2026-03-07-000004_AddJamPulangStandard.php
@@ -20,6 +20,8 @@ class AddJamPulangStandard extends Migration
 
     public function down()
     {
-        $this->forge->dropColumn('general_settings', 'jam_pulang_standard');
+        if ($this->db->fieldExists('jam_pulang_standard', 'general_settings')) {
+            $this->forge->dropColumn('general_settings', 'jam_pulang_standard');
+        }
     }
 }
